requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langgraph>=0.0.20
crewai>=0.28.0
openai>=1.0.0
ipython>=8.0.0
typing-extensions>=4.0.0
pinecone-client
requests
Pillow
open-clip-torch
torch 